rename sidh._math to sidh.math
