+++ sidh/math.py
@@ -20,7 +20,7 @@ def _try_composite(a, d, n, s):
 def is_prime(n):
     """
     Miller-Rabin primality test.
- 
+
     A return value of False means n is certainly not prime. A return value of
     True means n is very likely a prime.
     """
